requirements.txt
Flask==2.2.2
requests==2.28.1
aiohttp==3.8.3
websockets==10.4
websocket-client==1.4.2
grpcio-tools==1.51.1